requirements.txt
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
dateutils==0.6.12
dnspython==2.5.0
et-xmlfile==1.1.0
fake-headers==1.0.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
idna==3.6
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
Proxy-List-Scrapper==0.2.2
psycopg2==2.9.9
pyaes==1.6.1
pyasn1==0.5.1
pyenchant==1.6.6
pymongo==4.6.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
requests==2.31.0
rsa==4.9
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sql==2022.4.0
Telethon==1.34.0
trio==0.24.0
trio-websocket==0.11.1
tweety-ns==1.0.9.5
typing_extensions==4.9.0
urllib3==2.2.0
webencodings==0.5.1
websockets==12.0
wsproto==1.2.0
